authentication for normal and demo mode, update tests
please check https://github.com/iobis/smalldata/wiki/Authentication-and-authorization
to authenticate properly. For demo/testing, one can add the header
Authorization: Basic verysecret
to every call. This only works on servers running in DEMO mode,
and with the demokey set to 'verysecret' (as is on smalldata.obis.org
for now).